backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
supabase==2.7.4
openpyxl==3.1.5
python-dotenv==1.0.1
